FilesystemStorage without actor, but Arc<RwLock<>> instead
